+++ src/main/java/org/jarling/v2/models/accounts/CurrencyAndAmount.java
@@ -2,11 +2,13 @@ package org.jarling.v2.models.accounts;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 /**
  * Currency and amount
  */
 @Data
 public class CurrencyAndAmount {
     private final String currency;
-    private final long minorUnits;
+    private final BigInteger minorUnits;
 }
